Updated entity due date and group leaving date check

+++ discussion.php
@@ -18,10 +18,27 @@ if(!isset($_SERVER['HTTP_REFERER'])){
 <?php
 // Get topic name
 $data = $link->query("SELECT title FROM forum_topics WHERE topic_id=" . $_SESSION['topic_id']);
-if($data -> num_rows>0){
+if($data->num_rows>0){
     $topic_data = $data->fetch_assoc();
     $title = $topic_data['title'];
     echo $title;
+}
+
+// Check whether due date is passed or group enrolment is over
+$data = $link->query("SELECT due_date, DATE(SYSDATE()) FROM marked_entities WHERE marked_entity_id=" . $_SESSION['entity_id']);
+if ($data->num_rows>0){
+    $entity_data = $data->fetch_assoc();
+    $due_date = $entity_data['due_date'];
+    $current_date = $entity_data['DATE(SYSDATE())'];
+    $data = $link->query("SELECT left_group_date FROM group_users WHERE group_id=" . $_SESSION['group_id']);
+    if ($data->num_rows>0){
+        $group_data = $data->fetch_assoc();
+        $left_date = $group_data['left_group_date'];
+    }
+}
+$readonly = false;
+if($due_date < $current_date || isset($left_date)){
+    $readonly = true;
 }
 ?>
 </h1>
@@ -68,7 +85,10 @@ if($data -> num_rows>0){
         echo "<tr><td><pre>$text</pre></td>";
         echo "<td>";
         $f_text = addslashes($text);
-        echo "<input type='button' value='Reply' onclick=\"insertQuote('$f_text')\">";
+        // Display reply button for everyone
+        if(!$readonly){
+            echo "<input type='button' value='Reply' onclick=\"insertQuote('$f_text')\">";
+        }
         // Display delete button for admin or instructor
         if($_SESSION['role_id']<3){
             echo "<form class='form-button' method=post action='includes/delete_reply.php'>";
@@ -77,17 +97,22 @@ if($data -> num_rows>0){
         }
         // Display delete and edit button for original poster
         else{
-            if($user_id == $_SESSION['id']){
-                echo "<form class='form-button' method=post action='includes/delete_reply.php'>";
-                echo "<button type='submit' name='delete' value=$reply_id>Delete</button>";
-                echo "</form>";
+            if(!$readonly){
+                if($user_id == $_SESSION['id']){
+                    echo "<form class='form-button' method=post action='includes/delete_reply.php'>";
+                    echo "<button type='submit' name='delete' value=$reply_id>Delete</button>";
+                    echo "</form>";
+                }
             }
         }
-        if($user_id == $_SESSION['id']){
-            echo "<form class='form-button' method=post action='edit_reply.php'>";
-            echo "<input type='hidden' name='text' value='$text'>";
-            echo "<button name='edit' value=$reply_id>Edit</button>";
-            echo "</form>";
+
+        if(!$readonly){
+            if($user_id == $_SESSION['id']){
+                echo "<form class='form-button' method=post action='edit_reply.php'>";
+                echo "<input type='hidden' name='text' value='$text'>";
+                echo "<button name='edit' value=$reply_id>Edit</button>";
+                echo "</form>";
+            }
         }
         echo "</td></tr>";
         echo "</tbody></table>";
@@ -98,12 +123,14 @@ if($data -> num_rows>0){
 
 <p></p>
 
+<?php if(!$readonly){ ?>
 <!-- Reply textbox -->
 <form method=post action="includes/do_add_reply.php">
 <p><strong>Reply to topic:</strong><br>
 <textarea id='reply-box' name='reply' rows=8 cols=40 wrap=virtual></textarea></p>
 <p><button type="submit" name="submit">Reply</button></p>
 </form>
+<?php } ?>
 
 <script type="text/javascript">
     function insertQuote(text){
